pyproject.toml
[project]
name = "p1-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "sqlalchemy",
    "asyncpg",
]
